Threaded vmdr.get_host_list!!!
